Refactor: Move `image` to the `utils` directory
