+++ pkg/utils/image/image_test.go
@@ -12,7 +12,7 @@ import (
 	"k8s.io/utils/ptr"
 
 	"github.com/gardener/gardener-extension-image-rewriter/pkg/apis/config/v1alpha1"
-	. "github.com/gardener/gardener-extension-image-rewriter/pkg/image"
+	. "github.com/gardener/gardener-extension-image-rewriter/pkg/utils/image"
 )
 
 var _ = Describe("Image", func() {
