+++ pkg/webhook/pod/mutator_test.go
@@ -16,7 +16,7 @@ import (
 	"k8s.io/utils/ptr"
 
 	"github.com/gardener/gardener-extension-image-rewriter/pkg/apis/config/v1alpha1"
-	"github.com/gardener/gardener-extension-image-rewriter/pkg/image"
+	"github.com/gardener/gardener-extension-image-rewriter/pkg/utils/image"
 	. "github.com/gardener/gardener-extension-image-rewriter/pkg/webhook/pod"
 )
